// src/ArquivoMate2.Ui/src/app/client/models/save-email-settings-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EmailProviderType } from '../models/email-provider-type';
export interface SaveEmailSettingsRequest {
  autoReconnect?: boolean | null;
  connectionTimeout?: number | null;
  defaultFolder?: string | null;
  displayName?: string | null;
  password: string;
  port: number;
  providerType: EmailProviderType;
  server: string;
  useSsl?: boolean;
  username: string;
}