// src/ArquivoMate2.Ui/src/app/client/models/email-criteria-dto-api-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EmailCriteriaDto } from '../models/email-criteria-dto';
export interface EmailCriteriaDtoApiResponse {
  data?: EmailCriteriaDto;
  errorCode?: string | null;
  errors?: ({
[key: string]: Array<string>;
}) | null;
  message?: string | null;
  success?: boolean;
  timestamp?: string;
}