// src/ArquivoMate2.Ui/src/app/client/models/int-32-api-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Int32ApiResponse {
  data?: number;
  errorCode?: string | null;
  errors?: ({
[key: string]: Array<string>;
}) | null;
  message?: string | null;
  success?: boolean;
  timestamp?: string;
}